requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
aiosqlite==0.18.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
requests==2.31.0
